build: Switch to unified CMake build system
- Replace monolithic CMakeLists.txt and VS solution with modular CMake.
- Remove all .vcxproj/.vcxproj.filters/sln.
- Auto-detect CUDA architectures.
- Find or FetchContent zlib and protobuf via CMake.
- Generate version info via CMake.
- Fix compiler warnings.
- Update Docker scripts.
- Bump min CMake to 3.23.

+++ QTDialogs/SampleAnalyzerTab/SampleAnalyzerTab.cpp
@@ -407,9 +407,9 @@ void CSampleAnalyzerTab::ExportData()
 			int nOffsetDiameter = (dDiameter - m_dPSDStart)/dIntervalSize;
 			int nOffsetVel = (dVelocity - m_dVelStart)/dIntervalVel;
 			if ( nOffsetVel < 0 ) nOffsetVel = 0;
-			if ( nOffsetVel >= velDistribution.size() ) nOffsetVel = (int)velDistribution.size()-1;
+			if ( nOffsetVel >= (int)velDistribution.size() ) nOffsetVel = (int)velDistribution.size()-1;
 			if ( nOffsetDiameter < 0 ) nOffsetDiameter = 0;
-			if ( nOffsetDiameter >= sizeDistribution.size() ) nOffsetDiameter = (int)sizeDistribution.size()-1;
+			if ( nOffsetDiameter >= (int)sizeDistribution.size() ) nOffsetDiameter = (int)sizeDistribution.size()-1;
 			velDistribution[ nOffsetVel ] ++;
 			sizeDistribution[ nOffsetDiameter ] ++;
 		}
